Changed indentation
Changed the indentation of the documentation to be consistent with the other tickets and the developer's guide. Also for the outputs some of the secondary lines were aligned with the dash, while others were aligned with the first letter of the top line. Changed all to be aligned with the first letter (as mentioned in the paper).